clean up authorization failures

+++ src/TopoMojo.Api/Features/Dispatch/DispatchController.cs
@@ -38,16 +38,16 @@ namespace TopoMojo.Api.Controllers
         /// <returns></returns>
         [HttpPost("api/dispatch")]
         [SwaggerOperation(OperationId = "CreateDispatch")]
-        public async Task<Dispatch> Create([FromBody] NewDispatch model)
+        public async Task<ActionResult<Dispatch>> Create([FromBody] NewDispatch model)
         {
             await Validate(model);
 
-            AuthorizeAny(
+            if (!AuthorizeAny(
                 () => Actor.IsAdmin,
                 () => model.TargetGroup == Actor.Id, // gamespace agent
                 () => GamespaceService.CanManage(model.TargetGroup, Actor.Id).Result,
                 () => Actor.IsObserver && GamespaceService.HasValidUserScopeGamespace(model.TargetGroup, Actor.Scope).Result
-            );
+            )) return Forbid();
 
             var result = await DispatchService.Create(model);
 
@@ -63,18 +63,18 @@ namespace TopoMojo.Api.Controllers
         /// <returns></returns>
         [HttpGet("api/dispatch/{id}")]
         [SwaggerOperation(OperationId = "GetDispatch")]
-        public async Task<Dispatch> Retrieve([FromRoute] string id)
+        public async Task<ActionResult<Dispatch>> Retrieve([FromRoute] string id)
         {
             await Validate(new Entity { Id = id });
 
             var model = await DispatchService.Retrieve(id);
 
-            AuthorizeAny(
+            if (!AuthorizeAny(
                 () => Actor.IsAdmin,
                 () => model.TargetGroup == Actor.Id, // gamespace agent
                 () => GamespaceService.CanManage(model.TargetGroup, Actor.Id).Result,
                 () => Actor.IsObserver && GamespaceService.HasValidUserScopeGamespace(model.TargetGroup, Actor.Scope).Result
-            );
+            )) return Forbid();
 
             return await DispatchService.Retrieve(id);
         }
@@ -86,18 +86,18 @@ namespace TopoMojo.Api.Controllers
         /// <returns></returns>
         [HttpPut("api/dispatch")]
         [SwaggerOperation(OperationId = "UpdateDispatch")]
-        public async Task Update([FromBody] ChangedDispatch model)
+        public async Task<ActionResult> Update([FromBody] ChangedDispatch model)
         {
             await Validate(model);
 
             var dispatch = await DispatchService.Retrieve(model.Id);
 
-            AuthorizeAny(
+            if (!AuthorizeAny(
                 () => Actor.IsAdmin,
                 () => dispatch.TargetGroup == Actor.Id, // gamespace agent
                 () => GamespaceService.CanManage(dispatch.TargetGroup, Actor.Id).Result,
                 () => Actor.IsObserver && GamespaceService.HasValidUserScopeGamespace(dispatch.TargetGroup, Actor.Scope).Result
-            );
+            )) return Forbid();
 
             dispatch = await DispatchService.Update(model);
 
@@ -105,6 +105,7 @@ namespace TopoMojo.Api.Controllers
                 dispatch,
                 "UPDATE"
             );
+            return Ok();
         }
 
         /// <summary>
@@ -114,22 +115,23 @@ namespace TopoMojo.Api.Controllers
         /// <returns></returns>
         [HttpDelete("/api/dispatch/{id}")]
         [SwaggerOperation(OperationId = "DeleteDispatch")]
-        public async Task Delete([FromRoute] string id)
+        public async Task<ActionResult> Delete([FromRoute] string id)
         {
             await Validate(new Entity { Id = id });
 
             var entity = await DispatchService.Retrieve(id);
 
-            AuthorizeAny(
+            if (!AuthorizeAny(
                 () => Actor.IsAdmin,
                 () => entity.TargetGroup == Actor.Id, // gamespace agent
                 () => GamespaceService.CanManage(entity.TargetGroup, Actor.Id).Result,
                 () => Actor.IsObserver && GamespaceService.HasValidUserScopeGamespace(entity.TargetGroup, Actor.Scope).Result
-            );
+            )) return Forbid();
 
             await DispatchService.Delete(id);
 
             SendBroadcast(new Dispatch { Id = id, TargetGroup = entity.TargetGroup }, "DELETE");
+            return Ok();
         }
 
         /// <summary>
@@ -140,16 +142,16 @@ namespace TopoMojo.Api.Controllers
         /// <returns></returns>
         [HttpGet("/api/dispatches")]
         [SwaggerOperation(OperationId = "ListDispatches")]
-        public async Task<Dispatch[]> List([FromQuery] DispatchSearch model, CancellationToken ct)
+        public async Task<ActionResult<Dispatch[]>> List([FromQuery] DispatchSearch model, CancellationToken ct)
         {
             await Validate(model);
 
-            AuthorizeAny(
+            if (!AuthorizeAny(
                 () => Actor.IsAdmin,
                 () => model.gs == Actor.Id, // gamespace agent
                 () => GamespaceService.CanManage(model.gs, Actor.Id).Result,
                 () => Actor.IsObserver && GamespaceService.HasValidUserScopeGamespace(model.gs, Actor.Scope).Result
-            );
+            )) return Forbid();
 
             return await DispatchService.List(model, ct);
         }
